+++ Editor/Scripts/GLTFImporter.cs
@@ -23,12 +23,12 @@ using System.Linq;
 using System;
 using Object = UnityEngine.Object;
 using UnityGLTF.Loader;
-using GLTF.Schema;
 using GLTF;
 using UnityGLTF.Extensions;
 using UnityGLTF.Plugins;
 #if UNITY_2020_2_OR_NEWER
 using UnityEditor.AssetImporters;
+using UnityEditorInternal;
 using UnityEngine.Rendering;
 #else
 using UnityEditor.Experimental.AssetImporters;
@@ -47,7 +47,7 @@ namespace UnityGLTF
 #else
 	[ScriptedImporter(ImporterVersion, new[] { "glb" })]
 #endif
-    public class GLTFImporter : ScriptedImporter, IGLTFImportRemap
+    public class GLTFImporter : ScriptedImporter
     {
 	    private const int ImporterVersion = 9;
 
@@ -123,14 +123,19 @@ namespace UnityGLTF
         [NonReorderable] [SerializeField] private List<TextureInfo> _textures;
         [SerializeField] internal string _mainAssetIdentifier;
 
-        // TODO make internal and allow access for relevant assemblies
         internal List<TextureInfo> Textures => _textures;
 
+        // Import Plugin Overrides
+        [SerializeField] // but we can serialize their YAML representations as string
+        internal List<PluginInfo> _importPlugins = new List<PluginInfo>();
+        
         [Serializable]
         internal class ExtensionInfo
         {
 	        public string name;
+	        [HideInInspector]
 	        public bool supported;
+	        [HideInInspector]
 	        public bool used;
 	        public bool required;
         }
@@ -142,6 +147,15 @@ namespace UnityGLTF
 	        Fast = 0,
 	        Normal = 50,
 	        Best = 100
+		}
+
+		[Serializable]
+		public class PluginInfo
+		{
+			public string typeName;
+			public bool overrideEnabled;
+			public bool enabled;
+			public string jsonSettings;
 		}
 
         [Serializable]
@@ -229,22 +243,20 @@ namespace UnityGLTF
 
         public override void OnImportAsset(AssetImportContext ctx)
         {
-	        var plugins = new List<GltfImportPluginContext>();
-	        var context = new GLTFImportContext(ctx, plugins);
-	        var settings = GLTFSettings.GetOrCreateSettings();
-	        foreach (var plugin in settings.ImportPlugins)
+	        var settings = GLTFSettings.GetDefaultSettings();
+	        
+	        // make a copy, and apply import override settings
+	        foreach (var importPlugin in _importPlugins)
 	        {
-		        if (plugin != null && plugin.Enabled)
+		        if (importPlugin == null || !importPlugin.overrideEnabled) continue;
+		        var existing = settings.ImportPlugins.Find(x => x.GetType().FullName == importPlugin.typeName);
+		        if (existing)
 		        {
-			        var instance = plugin.CreateInstance(context);
-			        if(instance != null) plugins.Add(instance);
+			        existing.Enabled = importPlugin.enabled;
+			        JsonUtility.FromJsonOverwrite(importPlugin.jsonSettings, existing);
 		        }
 	        }
-
-	        foreach (var plugin in plugins)
-	        {
-		        plugin.OnBeforeImport();
-	        }
+	        var context = new GLTFImportContext(ctx, settings);
 
             GameObject gltfScene = null;
             AnimationClip[] animations = null;