+++ Runtime/Scripts/GLTFSceneImporter.cs
@@ -47,9 +47,9 @@ namespace UnityGLTF
 		public bool ImportBlendShapeNames = true;
 
 #if UNITY_EDITOR
-		public GLTFImportContext ImportContext = new GLTFImportContext(null, new List<GltfImportPluginContext>());
+		public GLTFImportContext ImportContext = new GLTFImportContext(null, GLTFSettings.GetOrCreateSettings());
 #else
-		public GLTFImportContext ImportContext = new GLTFImportContext(new List<GltfImportPluginContext>());
+		public GLTFImportContext ImportContext = new GLTFImportContext(GLTFSettings.GetOrCreateSettings());
 #endif
 
 		[NonSerialized]
@@ -360,6 +360,12 @@ namespace UnityGLTF
 
 					_isRunning = true;
 				}
+				
+				// TODO check where the right place is to call OnBeforeImport as early as possible
+				foreach (var plugin in Context.Plugins)
+				{
+					plugin.OnBeforeImport();
+				}
 
 				if (_options.ThrowOnLowMemory)
 				{
@@ -404,7 +410,10 @@ namespace UnityGLTF
 				}
 
 #if HAVE_MESHOPT_DECOMPRESS
-				await MeshOptDecodeBuffer(_gltfRoot);
+				if (Context.TryGetPlugin<MeshoptImportContext>(out _))
+				{
+					await MeshOptDecodeBuffer(_gltfRoot);
+				}
 #endif
 				await _LoadScene(sceneIndex, showSceneObj, cancellationToken);