+++ Runtime/Scripts/Plugins/UnlitMaterialsImport.cs
@@ -1,0 +1,14 @@
+namespace UnityGLTF.Plugins
+{
+    public class UnlitMaterialsImport: GltfImportPlugin
+    {
+        public override string DisplayName => "KHR_materials_unlit";
+        public override string Description => "Imports unlit materials.";
+        public override bool AlwaysEnabled => true;
+        public override GltfImportPluginContext CreateInstance(GLTFImportContext context)
+        {
+            // always enabled
+            return null;
+        }
+    }
+}