+++ Runtime/Scripts/SceneExporter/ExporterAnimationPointer.cs
@@ -1,10 +1,12 @@
 using System;
 using System.Collections.Generic;
+using System.Linq;
 using GLTF.Schema;
 using GLTF.Schema.KHR_lights_punctual;
 using UnityEngine;
 using UnityGLTF.Extensions;
 using UnityGLTF.JsonPointer;
+using UnityGLTF.Plugins;
 using Object = UnityEngine.Object;
 
 namespace UnityGLTF
@@ -383,7 +385,7 @@ namespace UnityGLTF
 					// Debug.LogWarning($"Implicitly handling animated property \"{propertyName}\" for target {animatedObject}", animatedObject);
 
 					// filtering for what to include / what not to include based on whether its target can be resolved
-					if (settings.UseAnimationPointer && animatedObject is Component _)
+					if (UseAnimationPointer && animatedObject is Component _)
 					{
 						var couldResolve = false;
 						var prop = $"/nodes/{channelTargetId}/{propertyName}";
@@ -527,16 +529,35 @@ namespace UnityGLTF
 						var colors = new Color[values.Length];
 						var strengths = new float[values.Length];
 						actuallyNeedSecondSampler = false;
-						for (int i = 0; i < values.Length; i++)
+						var pluginSettings = (_plugins.FirstOrDefault(x => x is MaterialExtensionsExportContext) as MaterialExtensionsExportContext)?.settings;
+						var emissiveStrengthSupported = pluginSettings != null && pluginSettings.KHR_materials_emissive_strength;
+						if (emissiveStrengthSupported)
 						{
-							DecomposeEmissionColor((Color) values[i], out var color, out var intensity);
-							colors[i] = color;
-							strengths[i] = intensity;
-							if (intensity > 1)
-								actuallyNeedSecondSampler = true;
+							for (int i = 0; i < values.Length; i++)
+							{
+								DecomposeEmissionColor((Color) values[i], out var color, out var intensity);
+								colors[i] = color;
+								strengths[i] = intensity;
+								if (intensity > 1)
+									actuallyNeedSecondSampler = true;
+							}
 						}
+						else
+						{
+							// clamp 0..1
+							for (int i = 0; i < values.Length; i++)
+							{
+								var c = (Color) values[i];
+								if (c.r > 1) c.r = 1;
+								if (c.g > 1) c.g = 1;
+								if (c.b > 1) c.b = 1;
+								colors[i] = c;
+							}
+						}
+						
 						Tsampler.Output = ExportAccessor(colors, false);
-						Tsampler2.Output = ExportAccessor(strengths);
+						if (emissiveStrengthSupported)
+							Tsampler2.Output = ExportAccessor(strengths);
 					}
 					else
 					{
@@ -569,7 +590,7 @@ namespace UnityGLTF
 			animation.Samplers.Add(Tsampler);
 			animation.Channels.Add(Tchannel);
 
-			if (settings.UseAnimationPointer)
+			if (UseAnimationPointer)
 				ConvertToAnimationPointer(animatedObject, propertyName, TchannelTarget);
 
 			// in some cases, extensions aren't required even when we think they might, e.g. for emission color animation.
@@ -606,7 +627,7 @@ namespace UnityGLTF
 				animation.Samplers.Add(Tsampler2);
 				animation.Channels.Add(Tchannel2);
 
-				if (settings.UseAnimationPointer)
+				if (UseAnimationPointer)
 					ConvertToAnimationPointer(animatedObject, secondPropertyName, TchannelTarget2);
 			}
 		}